Move check_input_coordinates from utils::internal to utils. Rename/move grb::internal::is_input_iterator to grb::utils::is_alp_matrix_iterator. Rename/move grb::internal::iterator_has_value_method to grb::utils::has_value_method. The latter two are now defined in graphblas/utils/iterators/type_traits.hpp instead of graphblas/type_traits.hpp. The former is retained in graphblas/utils/iterators/utils.hpp.
