+++ include/graphblas/type_traits.hpp
@@ -155,129 +155,6 @@ namespace grb {
 			static const constexpr bool value = false;
 		};
 
-		/**
-		 * Used to gauge whether a given type is an ALP matrix iterator.
-		 *
-		 * @tparam MatrixValType Value type of the matrix; if void, does not check for
-		 *                       the presence of a v() method that returns a nonzero
-		 *                       value.
-		 * @tparam IterT         The iterator type
-		 *
-		 * An ALP matrix iterator has the following methods:
-		 *  -# i(),
-		 *  -# j(), and
-		 *  -# v(), iff #MatrixValType is not void
-		 *
-		 * This iterator is only used internally, within ALP. It will never materialise
-		 * as a requirement on iterators in user-space.
-		 */
-		template< typename MatrixValType, typename IterT >
-		class is_input_iterator {
-
-			private:
-
-				// helper functions for determining, by return type, whether i, j, and v are
-				// defined
-
-				template< typename U >
-				static typename std::decay<
-					decltype(std::declval< U >().i())
-				>::type match_i(
-					typename std::decay< decltype(std::declval< U >().i()) >::type*
-				) {
-					return std::declval< U >().i();
-				}
-
-				template< typename U >
-				static void match_i( ... ) {}
-
-				template< typename U >
-				static typename std::decay<
-					decltype(std::declval< U >().j())
-				>::type match_j(
-					typename std::decay< decltype(std::declval< U >().j()) >::type*
-				) {
-					return std::declval< U >().j();
-				}
-
-				template< typename U >
-				static void match_j( ... ) {}
-
-				template< typename U >
-				static typename std::decay<
-					decltype(std::declval< U >().v())
-				>::type match_v(
-					typename std::decay< decltype(std::declval< U >().v()) >::type*
-				) {
-					return std::declval< U >().v();
-				}
-
-				template< typename U >
-				static void match_v( ... ) {}
-
-
-			public:
-
-				/** Type of the row index */
-				using RowIndexType = decltype( match_i< IterT >( nullptr ) );
-
-				/** Type of the column index */
-				using ColumnIndexType = decltype( match_j< IterT >( nullptr ) );
-
-				/** Type of the nonzero value */
-				using ValueType = decltype( match_v< IterT >( nullptr ) );
-
-				/**
-				 * Whether #IterT is an ALP matrix iterator
-				 */
-				static constexpr bool value =
-					!std::is_same< RowIndexType, void >::value &&
-					!std::is_same< ColumnIndexType, void >::value &&
-					std::is_integral< RowIndexType >::value &&
-					std::is_integral< ColumnIndexType >::value &&
-					(
-						std::is_same< MatrixValType, void >::value ||
-						!std::is_same< ValueType, void >::value
-					);
-
-		};
-
-		/**
-		 * Checks whether a given iterator type has a .v() method.
-		 *
-		 * @tparam T the iterator type
-		 *
-		 * I.e., whether \a T can be used for ingesting into a value (non-void) ALP
-		 * matrix.
-		 */
-		template< typename T >
-		class iterator_has_value_method {
-
-			private:
-
-				struct big { char a,b; };
-				template< typename U >
-				static typename std::decay< decltype( std::declval< U >().v() ) >::type	match(
-					typename std::decay< decltype( std::declval< U >().v() ) >::type*
-				) {
-					return std::declval< U >().v();
-				}
-
-				template< typename U >
-				static void match( ... ) {}
-
-
-			public:
-
-				/**
-				 * Whether \a T defines the .v() method.
-				 */
-				static constexpr bool value = !std::is_same<
-					decltype( match< T >( nullptr ) ), void
-				>::value;
-
-		};
-
 	} // end namespace grb::internal
 
 } // namespace grb
